Luke: Added and changed maps
- Added a map copy of posts_join_comments to dataset_by_id to the posts
  profile
- Added a map stream_join_granule and dataset_by_id to datasets profile.

In the future, all datasets, posts, science data etc will have a default
map whereby the key is a tuple consisting of the key identifier followed
by a type identifier usually an integer. This allows maps and reduces to
be executed quickly. Performance gains from this pattern are an
absolute must, the nomenclature dataset_by_id is just a logical
compliment to this consequence of events.