Fix test_interceptor_fails_when_send_multiple_messages
The test aims to test that a monitor can detect that sending
2 consequtive request messages with the same conversation-id is wrong.
Thus, the send method should be patched in a way that the request message pass
through the (outgoing) interceptor stack twice.
Therefore, patching BidirectionalChannel._send is wrong, because channel send is
positioned too late in the interceptor stack.
BidirectionalEndpointUnit._send should be patched instead.