Reomves filesystem cleanup from tests using --with-pycc
When the --with-pycc plugin is utilized a hosting container is launched
that hosts all the services and processes. This container's filesystem
is initialized and populated with information about datasets and other
system components.  When the container is done launching it signals the
nose process to go ahead and launch the container that plays host to the
tests.  This container unfortunately was deleting the filesystem!

This patches the filesystem force cleaning so that the container that
hosts the tests doesn't override and delete the parent container's file-
system.